requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
boto3==1.34.131
networkx==3.3
groq==0.9.0
python-dotenv==1.0.1
pydantic==2.7.4
